% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wts.R
\name{feat_wts}
\alias{feat_wts}
\title{Title Feature space weights generator}
\usage{
feat_wts(ref_df, obs_df, vars, probs)
}
\arguments{
\item{ref_df}{data.frame of pre-extracted raster values sampled from the broader data sets to serve as a reference for
weighting the observed samples.}

\item{obs_df}{data.frame of observed field samples with pre-extracted values of the rasters included in the reference dataset}

\item{vars}{character vector of the variable names being used for the weights}

\item{probs}{list of probability breaks used for each variable in the weighting}
}
\value{

}
\description{
Function to weight a sample set to better match a reference set of raster environmental variables
for use in model case weights of in weighted sub-sampling schemes. The function is currently set up to run with specific
variables from Ramcharan et al., 2018 SSSAJ. https://doi.org/10.2136/sssaj2017.04.0122.
Rasters at https://nrcs.box.com/s/p3dcg9lyw6ocmmlhtobm04ticax4thj5 as of 5/3/2021.
}
