% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRPI_estim_ranger.R
\name{gRPI_estim_ranger}
\alias{gRPI_estim_ranger}
\title{Global relative prediction interval estimation for given sample sets over a chosen spatial prediction domain.}
\usage{
gRPI_estim_ranger(
  x,
  gsamp,
  fm,
  os = "windows",
  grid,
  train.params,
  nthreads = detectCores() - 1,
  casewts = "tot_wts"
)
}
\arguments{
\item{x}{data.frame of all potential training instances for model}

\item{gsamp}{a random sample of the prediction domain with covariates pre-extracted}

\item{fm}{Formula to use for random forest model building}

\item{os}{string. Operating system for implementation. Options are "windows" or "linux".}

\item{train.params}{List of training parameters for building random forest models.}

\item{nthreads}{Integer number of logical cores to use for parallelization of function.}

\item{casewts}{Character name of field with case weights values}

\item{griddf}{A data.frame with geographic 'geo' and source 'srce' field expanded in a factorial to select
sample subsets for x.}
}
\value{
data.frame with original fields along with CV predictions
}
\description{
Global relative prediction interval estimation for given sample sets over a chosen spatial prediction domain.
}
