% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_extr_parallel.R
\name{parPTextr}
\alias{parPTextr}
\title{Extraction of covariate values to pedon locations to generate a regression matrix.}
\usage{
parPTextr(sp, gridlist, os = "windows", nthreads = detectCores() - 1)
}
\arguments{
\item{sp}{SpatialPointsDataFrame object of all pedons}

\item{gridlist}{Vector list of full path names of all covariate grids. Can be created using list.files using pattern= and full.names=T.}

\item{os}{Operating system. Supports "windows" or "linux"}

\item{nthreads}{Number of logical processors to use in parallel functions}
}
\value{
a data.frame with same original fields as sp plus the extracted values of rasters as new columns
}
\description{
Query values of covariate rasters for the location of each pedon being used to train predictive models. It is implemented as a parallel list
apply to speed up implementation. The parallel implementation varies by operating system.
}
