% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatCVranger.R
\name{SpatCVranger}
\alias{SpatCVranger}
\title{Title Spatial K-fold cross validation of a ranger random forest model.}
\usage{
SpatCVranger(
  sp,
  nfolds = 10,
  fm,
  rast,
  resol,
  os = "windows",
  train.params,
  nthreads = detectCores() - 1
)
}
\arguments{
\item{sp}{SpatialPointsDataFrame of all training instances for model}

\item{nfolds}{integer Number of folds to use for validation}

\item{fm}{Formula to use for random forest model building}

\item{rast}{raster object to be used for builing spatial validation blocks. Must be in projected CRS in units of meters.}

\item{resol}{integer resolution in kilometers for spatial validation block size}

\item{os}{string. Operating system for implementation. Options are "windows" or "linux".}

\item{train.params}{List of training parameters for building random forest models.}

\item{nthreads}{Integer number of logical cores to use for parallelization of function.}
}
\value{
data.frame with original fields along with CV predictions
}
\description{
Title Spatial K-fold cross validation of a ranger random forest model.
}
