% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVranger.R
\name{CVranger}
\alias{CVranger}
\title{Title K-fold cross validation of a ranger random forest model.}
\usage{
CVranger(
  x,
  nfolds = 10,
  fm,
  os = "windows",
  train.params,
  nthreads = detectCores() - 1,
  casewts = "tot_wts"
)
}
\arguments{
\item{x}{data.frame of all training instances for model}

\item{nfolds}{integer Number of folds to use for validation}

\item{fm}{Formula to use for random forest model building}

\item{os}{string. Operating system for implementation. Options are "windows" or "linux".}

\item{train.params}{List of training parameters for building random forest models.}

\item{nthreads}{Integer number of logical cores to use for parallelization of function.}

\item{casewts}{Character name of field with case weights values}
}
\value{
data.frame with original fields along with CV predictions
}
\description{
Title K-fold cross validation of a ranger random forest model.
}
