% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_stats.R
\name{valmetrics}
\alias{valmetrics}
\title{Title Tabulate validation statistics}
\usage{
valmetrics(xlst, trans, varrange, prop, depth)
}
\arguments{
\item{xlst}{list of data.frames or a dataframe of cross validation predictions}

\item{trans}{character string data transformation type: "none", "log10", "log", or "sqrt"}

\item{varrange}{numeric range of values representing the quantile interval used for uncertainty prediction intervals}

\item{prop}{character string name of property being modeled}

\item{depth}{integer depth in cm for model being evaluated}
}
\value{

}
\description{
Title Tabulate validation statistics
}
